# Optimized existence checks