Optimized existence checks and object lookups.

*This branch:*
```
lookup static route x 29,182,417 ops/sec ±1.19% (87 runs sampled)
lookup dynamic route x 1,468,042 ops/sec ±1.28% (86 runs sampled)
lookup dynamic multi-parametric route x 819,827 ops/sec ±1.26% (90 runs sampled)
lookup dynamic multi-parametric route with regex x 792,738 ops/sec ±0.56% (92 runs sampled)
lookup long static route x 3,332,216 ops/sec ±1.05% (91 runs sampled)
find static route x 29,208,520 ops/sec ±1.01% (87 runs sampled)
find dynamic route x 1,560,609 ops/sec ±1.20% (83 runs sampled)
find dynamic multi-parametric route x 847,141 ops/sec ±1.07% (88 runs sampled)
find dynamic multi-parametric route with regex x 827,792 ops/sec ±0.82% (91 runs sampled)
find long static route x 4,145,462 ops/sec ±1.07% (89 runs sampled)
```

*Master:*
```
lookup static route x 27,668,776 ops/sec ±1.58% (84 runs sampled)
lookup dynamic route x 1,465,033 ops/sec ±1.20% (89 runs sampled)
lookup dynamic multi-parametric route x 811,878 ops/sec ±1.26% (87 runs sampled)
lookup dynamic multi-parametric route with regex x 776,741 ops/sec ±0.79% (90 runs sampled)
lookup long static route x 3,263,162 ops/sec ±1.16% (91 runs sampled)
find static route x 29,058,641 ops/sec ±1.06% (86 runs sampled)
find dynamic route x 1,574,922 ops/sec ±1.30% (89 runs sampled)
find dynamic multi-parametric route x 851,767 ops/sec ±1.12% (88 runs sampled)
find dynamic multi-parametric route with regex x 824,612 ops/sec ±0.84% (91 runs sampled)
find long static route x 4,165,586 ops/sec ±1.12% (90 runs sampled)
```